# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin OpenApi",
        "description": "this is a api documentation",
        "contact": {
            "name": "Emad shirzad",
            "email": "shemad2022@gmail.com"
        },
        "license": {
            "name": "GPL-v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/auth/envoy/register-user": {
            "post": {
                "tags": [
                    "Envoy"
                ],
                "summary": "register user by envoy",
                "description": "register envoy",
                "operationId": "register envoy",
                "requestBody": {
                    "description": "tasks input",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "username ",
                                        "type": "string",
                                        "example": "emad"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "password confirm",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnvoyModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/auth/envoy/envoy-users": {
            "get": {
                "tags": [
                    "Envoy"
                ],
                "summary": "users envoy list",
                "description": "envoy info",
                "operationId": "b3304048de557bd57536d3dd10c0e0ac",
                "responses": {
                    "200": {
                        "description": "Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "an 'unexpected' error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorModel": {
                "title": "Error Model",
                "description": "Error Model",
                "properties": {
                    "message": {
                        "description": "error message",
                        "type": "string",
                        "format": "",
                        "example": "my error message"
                    },
                    "errors": {
                        "description": "errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "error1",
                            "error2"
                        ]
                    }
                },
                "type": "object"
            },
...
            "EnvoyModel": {
                "title": "Envoy Model",
                "description": "Represents a envoy",
                "properties": {
                    "id": {
                        "description": "user ID",
                        "type": "integer",
                        "format": "int32"
                    },
                    "username": {
                        "description": "username",
                        "type": "string"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "password": {
                        "description": "password",
                        "type": "string"
                    },
                    "phone": {
                        "description": "phone envoy",
                        "type": "string"
                    },
                    "is_ban": {
                        "description": "is_ban",
                        "type": "boolean"
                    },
                    "address": {
                        "description": "address",
                        "type": "integer"
                    },
                    "balance": {
                        "description": "balance",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserModel": {
                "title": "User Model",
                "description": "Represents a user",
                "properties": {
                    "id": {
                        "description": "user ID",
                        "type": "integer",
                        "format": "int32"
                    },
                    "username": {
                        "description": "username",
                        "type": "string"
                    },
                    "password": {
                        "description": "password",
                        "type": "string"
                    },
                    "created_by": {
                        "description": "created by envoy",
                        "type": "string"
                    },
                    "is_ban": {
                        "description": "is_ban",
                        "type": "boolean"
                    },
                    "subscription_duration": {
                        "description": "subscription_duration",
                        "type": "integer"
                    },
                    "last_purchase_date": {
                        "description": "last purchase date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "created date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "description": "Login with username and password to get the authentication token <div>Example: Bearer token</div>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}